진행상황 백업: TODO: Backup 부분 코드 정리

+++ app/src/main/java/vh1981/com/funnyphotosstorage/ImageDatabaseManager.java
@@ -4,9 +4,11 @@ import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
+import android.graphics.Path;
 import android.util.DebugUtils;
 import android.util.Pair;
 
+import java.io.File;
 import java.util.ArrayList;
 import java.util.Set;
 
@@ -22,6 +24,10 @@ public class ImageDatabaseManager {
     private static Context _context;
     private static String[] _defaultTags = {"유머","진지","급정색", "근성", "간지"};
     private static long UN_TAGGED = -1;
+
+    ////////////////////////////////////////////////////////////////////////////////
+    /// 공개 기능들 :
+    ////////////////////////////////////////////////////////////////////////////////
     public static void setContext(Context context)
     {
         ImageDatabaseManager._context = context;
@@ -30,8 +36,6 @@ public class ImageDatabaseManager {
     private ImageDatabaseManager()
     {
         _dbHelper = new ImgDbHelper(_context);
-
-
         initDefaultTags();
     }
 
@@ -50,6 +54,9 @@ public class ImageDatabaseManager {
             ContentValues values = new ContentValues();
             values.put(ImgDbHelper.TagsEntry.COLUMN_NAME_TAGNAME, tag);
             long id = db.insert(ImgDbHelper.TagsEntry.TABLE_NAME, null, values);
+
+            File f = _context.getDatabasePath(_dbHelper.DATABASE_NAME);
+            String pathStr = f.getAbsolutePath();
             if (id < 0) {
                 assert false;
             }
@@ -61,26 +68,30 @@ public class ImageDatabaseManager {
         }
     }
 
-    public boolean removeTag(long tagId) {
-        // imageTag에서 해당 tagId를 -1로 변경
-        SQLiteDatabase db = _dbHelper.getWritableDatabase();
-        ContentValues values = new ContentValues();
-        values.put(ImgDbHelper.ImageTagEntry.COLUMN_NAME_TAGID, UN_TAGGED);
-        String whereClause = ImgDbHelper.ImageTagEntry.COLUMN_NAME_TAGID + " = ?";
-        String[] whereArgs = new String[]{Long.toString(tagId)};
-        int cnt = db.update(ImgDbHelper.ImageTagEntry.TABLE_NAME, values, whereClause, whereArgs);
-        DebugLog.TRACE("updated rows = " + Integer.toString(cnt));
-
-        // tag에서 삭제
-        whereClause = ImgDbHelper.TagsEntry._ID + " = ?";
-        whereArgs = new String[]{Long.toString(tagId)};
-        cnt = db.delete(ImgDbHelper.TagsEntry.TABLE_NAME, whereClause, whereArgs);
-        if (cnt == 0) {
-            DebugLog.TRACE("Error!");
-            assert false;
+    public boolean removeTag(String tag)
+    {
+        int tid = findTagId(tag);
+        if (tid >= 0) {
+            // tag에서 삭제
+            SQLiteDatabase db = _dbHelper.getWritableDatabase();
+            String whereClause = ImgDbHelper.TagsEntry._ID + " = ?";
+            String[] whereArgs = new String[]{Long.toString(tid)};
+            int cnt = db.delete(ImgDbHelper.TagsEntry.TABLE_NAME, whereClause, whereArgs);
+            if (cnt == 0) {
+                DebugLog.TRACE("Error!");
+                assert false;
+            }
+            return true;
         }
-
         return true;
+    }
+
+    public boolean tagExists(String tag)
+    {
+        if (findTagId(tag) >= 0) {
+            return true;
+        }
+        return false;
     }
 
     public void initDefaultTags() {
@@ -114,11 +125,35 @@ public class ImageDatabaseManager {
         return tags;
     }
 
-    /**
-     * Image테이블에 파일명을 추가한다.
-     * @param filepath : 파일 경로 + 이름
-     * @return : 실패시 -1, 성공시 rowId
-     */
+    ////////////////////////////////////////////////////////////////////////////////
+    /// 내부함수 :
+    ////////////////////////////////////////////////////////////////////////////////
+    private int findTagId(String tag) {
+        String query = "select " + ImgDbHelper.TagsEntry._ID + " from " + ImgDbHelper.TagsEntry.TABLE_NAME + " where "
+                + ImgDbHelper.TagsEntry.COLUMN_NAME_TAGNAME + " = " + "'" + tag + "'";
+        SQLiteDatabase db = _dbHelper.getWritableDatabase();
+        Cursor c = db.rawQuery(query, null);
+        int tagId = 0;
+        boolean found = false;
+        assert c != null;
+        if (c != null) {
+            if (c.moveToFirst()) {
+                do {
+                    tagId = c.getInt(c.getColumnIndex(ImgDbHelper.TagsEntry._ID));
+                    found = true;
+                } while (c.moveToNext());
+            }
+            c.close();
+        }
+
+        if (found == false) {
+            return -1;
+        }
+        return tagId;
+    }
+
+
+    /*
     public long createImage(String filepath)
     {
         SQLiteDatabase db = _dbHelper.getWritableDatabase();
@@ -165,30 +200,6 @@ public class ImageDatabaseManager {
         return true;
     }
 
-    public int findTagId(String tag) {
-        String query = "select " + ImgDbHelper.TagsEntry._ID + " from " + ImgDbHelper.TagsEntry.TABLE_NAME + " where "
-                + ImgDbHelper.TagsEntry.COLUMN_NAME_TAGNAME + " = " + "'" + tag + "'";
-        SQLiteDatabase db = _dbHelper.getWritableDatabase();
-        Cursor c = db.rawQuery(query, null);
-        int tagId = 0;
-        boolean found = false;
-        assert c != null;
-        if (c != null) {
-            if (c.moveToFirst()) {
-                do {
-                    tagId = c.getInt(c.getColumnIndex(ImgDbHelper.TagsEntry._ID));
-                    found = true;
-                } while (c.moveToNext());
-            }
-            c.close();
-        }
-
-        if (found == false) {
-            return -1;
-        }
-        return tagId;
-    }
-
     public String findTag(int tagId) {
         String query = "select " + ImgDbHelper.TagsEntry.COLUMN_NAME_TAGNAME + " from " + ImgDbHelper.TagsEntry.TABLE_NAME + " where "
                 + ImgDbHelper.TagsEntry._ID + " = " + tagId;
@@ -230,8 +241,10 @@ public class ImageDatabaseManager {
         }
         return filepath;
     }
+    */
 
-    public ImageContainer getImageContainer(int tagId, BitmapSupplier supplier /*Bitmap 공급자*/)
+    /*
+    public ImageContainer getImageContainer(int tagId, BitmapSupplier supplier)
     {
         ImageContainer imageContainer = new ImageContainer();
         String tag;
@@ -260,5 +273,13 @@ public class ImageDatabaseManager {
             c.close();
         }
         return imageContainer;
+    }
+    */
+
+    public String dbFilePath()
+    {
+        File f = _context.getDatabasePath(_dbHelper.DATABASE_NAME);
+        String pathStr = f.getAbsolutePath();
+        return pathStr;
     }
 }
